scanner/ScanningApp/requirements.txt
pillow
twain
